requirements.txt
viam-sdk==0.34.0